Implement a basic working game starting.
This commit primarily implements the start script schema and
a very basic server that can get http request with the start
specification and start a single instance of engine dedicated.

The main and game manager are entirely unfinished, minimal setup
to be able run launch anything but startscript generation part
is more or less in a good shape.