Introduce engine runner abstraction
The EngineRunner manages the lifecycle of the engine process, it
starts it, faciliates communication via autohost interface, and
monitors for failures and exit.

EngineRunner is one of the last building blocks that handle
interaction with the engine, on top of which we can now build the
business logic of host: multiplexing of game instances.