Replace embedded tachyon schema definition with official package
Now that the tachyon-protocol package contains a stable and proper
definitions of the autohost protocol based on that was in this repo,
this change rips out all custom schemas, and replaces it with the
types and validators that come from the official protocol package.

There are some static typing improvements with this change, there is
less casting required as more code is automatically generated in the
upstream package.